requirements.txt
Flask==2.2.5
flask-login==0.6.2
Flask-SQLAlchemy==3.0.3
Werkzeug==2.2.3
Pillow==10.0.0
torch==2.2.0
torchvision==0.17.0
numpy==1.26.0